menu controls backgrounds update
