--- core/src/org/academiadecodigo/murlogs/GameScreen.java
@@ -1,119 +1,0 @@
-package org.academiadecodigo.murlogs;
-
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Input.Keys;
-import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.Texture;
-
-public class GameScreen implements Screen {
-    final App game;
-
-    //images
-    Texture background;
-
-
-    //sound & Music
-
-
-    OrthographicCamera camera;
-    //players
-
-
-    public GameScreen(final App game) {
-        this.game = game;
-
-        // load the images
-        background = new Texture(Gdx.files.internal("fixbg.png"));
-
-
-        // load sound
-
-
-        // create the camera and the SpriteBatch
-        camera = new OrthographicCamera();
-        camera.setToOrtho(false, 1024, 576);
-
-
-        // create player
-
-
-
-    }
-
-
-    private void spawnRaindrop() {
-
-    }
-
-
-    @Override
-    public void render(float delta) {
-        // clear the screen with a dark blue color. The
-        // arguments to glClearColor are the red, green
-        // blue and alpha component in the range [0,1]
-        // of the color to be used to clear the screen.
-        Gdx.gl.glClearColor(0, 0, 0.2f, 1);
-        Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-
-        // tell the camera to update its matrices.
-        camera.update();
-
-        // tell the SpriteBatch to render in the
-        // coordinate system specified by the camera.
-        game.batch.setProjectionMatrix(camera.combined);
-
-
-        // begin a new batch and draw the player and others
-        game.batch.begin();
-        game.batch.draw(background, 0, 0);
-        game.batch.end();
-/*
-        // process user input
-        if (Gdx.input.isKeyPressed(Keys.LEFT))
-            player.x -= 200 * Gdx.graphics.getDeltaTime();
-        if (Gdx.input.isKeyPressed(Keys.RIGHT))
-            player.x += 200 * Gdx.graphics.getDeltaTime();
-
-        // make sure the bucket stays within the screen bounds
-        if (player.x < 0)
-            player.x = 0;
-        if (player.x > 1024 - 128)
-            player.x = 1024 - 128;
-
-*/
-    }
-
-    @Override
-    public void resize(int width, int height) {
-    }
-
-    @Override
-    public void show() {
-        // start the playback of the background music
-        // when the screen is shown
-        //music.play();
-    }
-
-    @Override
-    public void hide() {
-    }
-
-    @Override
-    public void pause() {
-    }
-
-    @Override
-    public void resume() {
-    }
-
-    @Override
-    public void dispose() {
-        //player.dispose();
-        //other.dispose();
-        //sound.dispose();
-        //music.dispose();
-    }
-
-}
